refactor(daemon): relocate statusz client and category mapping to internal/blocknode/shaper
Move the statusz REST client and the statusz-category→policy mapping/diff out of
the daemon's internal/daemon/blocknode into a new internal/blocknode/shaper
package, so the CLI can own statusz parsing and the daemon stops importing it.
Pure relocation (git mv, no logic change) plus:

- traffic_shaper_monitor.go: drop the live-set lister and the reconcilePolicies
  wrapper; runStatuszPoll becomes an idle stub. The daemon no longer parses
  statusz or reads nft in-process — that reconcile logic moves to the
  `block node reconcile-shaper` worker.

Part of #893.
