+++ internal/blocknode/shaper/policy_map.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package blocknode
+package shaper
 
 import (
 	"context"
